Don't munge anything other than HTML
This is a complete re-write of how the filtering works. It will be a
little slower, but should scale much better. Both in terms of
performance, and readability.

HTML is the only thing we are interested in. Before we were blindly
doing the same mungeing over the whole email. It could have had matches
in the headers, or attachments, which definitely should not be messed
with. We were seeing manifestations of it messing with the headers such
as From, and Subject appearing to go missing. Now those should be
entirely left alone.